# api_gateway/OpenAPI.yml
openapi: 3.0.3
info:
  title: StudyFlow API
  version: 1.0.0

  security:
    - telegramAuth: []
components:
  securitySchemes:
    telegramAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Two formats supported:

        **1. Telegram Mini App initData (production):**
        `tma <initData from Telegram.WebApp.initData>`
        - initData is validated per Telegram Bot API spec (HMAC-SHA256 with bot token)
        - No client-side secret needed
        - TTL: 24 hours from auth_date

        **2. Legacy HMAC (development only):**
        `telegram <telegramId>:<utcTimestamp>:<hmacHex>`
        - HMAC-SHA256 of `{telegramId}:{utcTimestamp}` with TELEGRAM_SECRET as key
        - ±5 minute clock skew
        - Disable in production with AUTH_DISABLE_LEGACY_HMAC=true
  schemas:
    UserRole:
      type: string
      enum:
        - tutor
        - student
    UserStatus:
      type: string
      enum:
        - active
        - deleted
    TutorStudentStatus:
      type: string
      enum:
        - invited
        - active
    User:
      type: object
      properties:
        id:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        authProvider:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        firstName:
          type: string
        lastName:
          type: string
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    UserPublic:
      type: object
      properties:
        id:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        firstName:
          type: string
        lastName:
          type: string
    TutorProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        paymentInfo:
          type: string
        lessonPriceRub:
          type: integer
        lessonConnectionLink:
          type: string
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    TutorStudent:
      type: object
      properties:
        id:
          type: string
        tutorId:
          type: string
        studentId:
          type: string
        lessonPriceRub:
          type: integer
        lessonConnectionLink:
          type: string
        status:
          $ref: '#/components/schemas/TutorStudentStatus'
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    InvitationStatus:
      type: string
      enum:
        - active
        - used
        - revoked
    Invitation:
      type: object
      properties:
        id:
          type: string
        tutorId:
          type: string
        token:
          type: string
          description: One-time UUID token used to build the student-facing invite link
        status:
          $ref: '#/components/schemas/InvitationStatus'
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      example:
        error: Not Found
...
paths:
...
  /users/users/{id}:
    get:
      summary: Get public user profile
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User public info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update user info
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                timezone:
                  type: string
                role:
                  type: string
                  enum:
                    - tutor
                    - student
      responses:
        '200':
          description: User updated
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Soft-delete user account
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted (status set to "deleted")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/users/invitations:
    post:
      summary: Create an invitation
      description: >
        Creates a one-time invitation for the current tutor. The returned
        `token` is used by the frontend to build a student-facing invite link.
        Requires the tutor role.
      operationId: createInvitation
      responses:
        '200':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (caller is not a tutor)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List the current tutor's invitations
      description: Returns all invitations created by the current tutor, newest first. Requires the tutor role.
      operationId: listInvitations
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invitation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (caller is not a tutor)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/users/invitations/{id}:
    delete:
      summary: Revoke an invitation
      description: Revokes an invitation owned by the current tutor (sets status to "revoked"). Requires the tutor role.
      operationId: revokeInvitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Invitation id
      responses:
        '200':
          description: Invitation revoked
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (invitation belongs to another tutor)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/users/invitations/{token}/accept:
    post:
      summary: Accept an invitation by token
      description: >
        A student accepts a tutor's invitation using its token. On success a
        tutor-student relationship is created with status "active" and the
        invitation is marked "used". Requires the student role.
      operationId: acceptInvitation
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Invitation token (UUID)
      responses:
        '200':
          description: Invitation accepted; tutor-student relationship created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorStudent'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (caller is not a student)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Token not found or invitation no longer active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-profiles/{user_id}:
    get:
      summary: Get tutor profile by user_id
      operationId: getTutorProfile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tutor profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorProfile'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update tutor profile
      operationId: updateTutorProfile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paymentInfo:
                  type: string
                lessonPriceRub:
                  type: integer
                lessonConnectionLink:
                  type: string
      responses:
        '200':
          description: Tutor profile updated
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students/by-tutor/{tutor_id}:
    get:
      summary: List tutor-student relationships by tutor
      operationId: listTutorStudents
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TutorStudent'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students/by-student/{student_id}:
    get:
      summary: List tutor-student relationships by student
      operationId: listTutorsForStudent
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TutorStudent'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students/{tutor_id}/{student_id}:
    get:
      summary: Get tutor-student relationship
      operationId: getTutorStudent
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Relationship found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorStudent'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update tutor-student relationship
      operationId: updateTutorStudent
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lessonPriceRub:
                  type: integer
                lessonConnectionLink:
                  type: string
                status:
                  $ref: '#/components/schemas/TutorStudentStatus'
      responses:
        '200':
          description: Relationship updated
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete tutor-student relationship
      operationId: deleteTutorStudent
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Relationship deleted
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students/{tutor_id}/accept:
    post:
      summary: Accept tutor invitation
      operationId: acceptInvitationFromTutor
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation accepted
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tutor or relationship not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/tutor-students:
    post:
      summary: Create tutor-student relationship
      operationId: createTutorStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tutorId:
                  type: string
                studentId:
                  type: string
                lessonPriceRub:
                  type: integer
                lessonConnectionLink:
                  type: string
              required:
                - tutor_id
                - student_id
      responses:
        '200':
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorStudent'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Relationship already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


  # files
...